requirements.txt
dash==2.18.2
numpy==2.2.0
pandas==2.2.3
pyMapVBVD==0.6.1